config/loader: Introduce a new config loading pkg
This package is supposed to be used from now on to load any configuration from
either global or local configuration file. It is architected so that any config
marshalling can be deferred into the packages that are interested in the
relevant configuration sections.

Any package wishing to load any config is supposed to implement ConfigSpec
interface, which is then used to get the global and local ConfigContainer
objects, which implement methods for handling both global and local
configuration. The interfaces are designed in a way that the package is able to
prompt the user in case there is any configuration missing basically on demand.
ConfigContainer interface contains methods for both loading and validations the
configuration as well as prompting the user, so it is as well used during `repo
bootstrap` to prompt the user for the relevant config for the given repository.

The usage will become more clear in the commits to come, which contain
implementations of these interfaces for concrete modules.

Change-Id: f28f91e262
Story-Id: SF-220