repo prune: Don't panic when story not found
In case the story that a commit is associated with is not found in the
issue tracker, repo prune panics, which is not really the good and
robust way how to handle things.

In case the story is not found in the issue tracker, we simply skip any
checks considering such a commit, basically the same as for the
unassigned commits.

Change-Id: 384b827cb4
Story-Id: SF-236