config/loader: Make empty local config work
Marshalling and unmarshalling of an empty local config section was not
symmetrical. While marshalling was skipping the section altogether,
unmarshalling was expecting a section of the given key to always exist.

The solution is to indeed create a config section filled with '{}' (en
empty object) during repo bootstrap, which fixed unmarshalling since the
config section for the given key exists now and no error is returned.

Change-Id: b618b5d9ee
Story-Id: SF-220