release deploy: Actually close the release
runningRelease.Release() was not being called at all during `release
deploy`, which is a major bug. The only luck was that the method is
empty for Pivotal Tracker, so nobody noticed anything.

This change makes sure the method is invoked during `release deploy`.
The relevant interface in the common package had to be changed to
include action.Action that can be used to roll back the changes.

Also, the existing modules were updated to match the new interface.

Change-Id: bd2bd9c468
Story-Id: SF-245