config/loader: Use emptyModuleConfigSpec internally
Fix the issue with requiring modules to use loader.NewEmptyModuleConfigSpec()
for ConfigSpec.LocalConfig() instead of just returning nil. From the interface
perspective, returning nil is perfectly fine and the loader package has to deal
with it internally.

This commit changes this requirement and loader package now uses the empty
module config spec implementation just internally and it is not exposed to the
user. Modules can safely return nil as the local config container.

Change-Id: 50505d3b11
Story-Id: SF-220