prompt/storyprompt: Improve the story dialog
The original motivation was to make it possible to filter the stories in the
dialog by title substring. As the result, the dialog is now configurable and
much more powerful, making it possible to implement custom dialog options, like
already mentioned filtering of stories by title substring and so on.

Change-Id: c930019690
Story-Id: salsaflow/salsaflow#86