changes: Handle bad commit messages better
In case the user skips the commit title and puts the commit message on that
line, the commit will be found as a story commit, but the parser will fail to
get the tag value, leading to panics.

This change makes SF detect these bad commits, print a warning, drop them and
continue, effectively preventing these panics from happening.

Change-Id: 4c3ad7c5f0
Story-Id: salsaflow/salsaflow#112