review post: Change the story branch push behavior
From now on, review post is no counting on the fact that the story branch must
be always pushed. In case -parent is specified, review post will autimatically
merge the story branch into the parent branch and it will push the parent branch
before posting commits for review to make sure the commits are present in the
remote repository.

In case -no_merge is specified, no merge occurs and SF makes sure the story
branch is up to date with the remote repository, which usually means that it is
pushed during review post.

There is also another new flag, -merge_no_ff, which can be used to make sure the
merge happening uses git merge --no-ff. There is merge dialog as there used to
be, so this flag is supposed to be used instead.

Change-Id: 39861fd8cd
Story-Id: salsaflow/salsaflow#84