release start: Check staging branch explicitly
The staging branch is checked whether it is up to date later in the process
inside release.PromptUserToConfirmStart(), which calls
releases.ListNewTrunkCommits(), which contains this checks. This change pulls
this check to the top of the release start process so that we fail earlier in
case the staging branch is not up to date.

Change-Id: 21b614aa26
Story-Id: salsaflow/salsaflow#131