review post: Rewrite as few commits as possible
When review post is passed a revision range by using -parent, it always
cherry-picks all the commits in the range. However, there is no need to amend
the commits that are at the beginning of the list that have the Story-Id tag
filled in. These are ready to be posted for review.

So, this commits makes review post more clever by leaving the leading commits
with the Story-Id tag filled in where they were. This actually fixes a bug with
some commits being posted for review multiple times. Since review post was
cherry-picking all commits, SHA hashes changed and all the commits were always
posted for review. This is now fixed, obviously.

Another way would be to use Change-Id tag to search for review requests.

Change-Id: 3cf0090010
Story-Id: salsaflow/salsaflow#78