Finalise code review release during release deploy
It used to be happening during release stage, but that actually doesn't make
much sense since the client can request some changes, which then need to be
reviewed. This will break e.g. because the review milestone for GitHub is closed
already. And in general, it really makes more sense to close the release in the
code review tool at the same time as in the issue tracker.

The code for initialising/finalising a release in the code review tool was split
into a separate common.Release interface, which also makes it possible to check
whether the release can be closed before trying to close it.

Also aligned common.IssueTracker interface to make more sense, e.g. the release
object constructors do not return errors any more since no module actually
returns anything but nil.

Change-Id: 73384cd71d
Story-Id: salsaflow/salsaflow#136