Implement a new command: cherry-pick
This command can be used to move chosen commits onto the release branch, but by
specifying a custom target branch, this command can be used to perform any
general cherry-pick. It is basically just a wrapper around git cherry-pick that
lets you stay on the branch where you are, which can save some annoying typing.

Change-Id: c92a9b439f
Story-Id: salsaflow/salsaflow#134